move BuildTestError to buildtest/exceptions.py and refactor import statement
