+++ tests/test_exceptions.py
@@ -1,11 +1,12 @@
 import pytest
 from buildtest.utils.command import BuildTestCommand
-from buildtest.utils.exceptions import BuildTestError
+from buildtest.exceptions import BuildTestError
+
 
 class TestBuildTestError:
-
     @pytest.mark.xfail(
-        reason="Testing to see if exception of type BuildTestError is raised", raises=BuildTestError
+        reason="Testing to see if exception of type BuildTestError is raised",
+        raises=BuildTestError,
     )
     def test_exception(self):
         """This test will check if we can raise Exception of type BuildTestError"""
@@ -15,4 +16,6 @@ class TestBuildTestError:
         a.execute(cmd)
         err = a.get_error()
         ret = a.returnCode()
-        raise BuildTestError(f"Command: {cmd} failed, Return Code: {ret} with Error: {err}")
+        raise BuildTestError(
+            f"Command: {cmd} failed, Return Code: {ret} with Error: {err}"
+        )
